TOMEE-1136 first step to keep context.stop() working, still some testing to validate
git-svn-id: https://svn.apache.org/repos/asf/tomee/tomee/trunk@1573276 13f79535-47bb-0310-9956-ffa450edef68

+++ tomee/tomee-catalina/src/main/java/org/apache/tomee/catalina/TomcatWebAppBuilder.java
@@ -718,12 +718,14 @@ public class TomcatWebAppBuilder implements WebAppBuilder, ContextListener, Pare
             if (contextInfo != null) {
                 final StandardContext standardContext = contextInfo.standardContext;
 
-                undeploy(standardContext, contextInfo);
-                final File extracted = Contexts.warPath(standardContext);
-                if (isExtracted(extracted)) {
-                    deleteDir(extracted);
+                if (!appInfo.webAppAlone || !appInfo.properties.containsKey("tomee.destroying")) {
+                    undeploy(standardContext, contextInfo);
+                    final File extracted = Contexts.warPath(standardContext);
+                    if (isExtracted(extracted)) {
+                        deleteDir(extracted);
+                    }
+                    removeContextInfo(standardContext);
                 }
-                removeContextInfo(standardContext);
             }
         }
     }
